tpcc: move adapter and types to shared dir
